Change to v1.0.8 and Ardunio_boards v1.0.5-2 - Fix branch check - Check if FW_FLAVAVOR and FW_FLAVERSION are correct
Fix typo

+++ PF-build.sh
@@ -180,6 +180,9 @@
 # 20 Jun 2022, 3d-gussner, Change to Ardunio_boards v1.0.5-1
 # 20 Jun 2022, wavexx    , New PO-based language translation support
 # 20 Jun 2022, 3d-gussner, fix Mk2.5/S zip after change PO-based language translation support
+# 06 Jul 2022, 3d-gussner, Change to v1.0.8 and Ardunio_boards v1.0.5-2
+# 06 Jul 2022, 3d-gussner, Fix branch check
+# 12 Jul 2022, 3d-gussner, Check if FW_FLAVAVOR and FW_FLAVERSION are correct
 
 SCRIPT_PATH="$( cd "$(dirname "$0")" ; pwd -P )"
 export SRCDIR=$SCRIPT_PATH
@@ -203,6 +206,7 @@ case "$1" in
     22) echo "$(tput setaf 1)PF-build.sh has been interrupted/failed. $(tput setaf 6)Restoring 'config.h'$(tput sgr0)" ; if [ $OUTPUT == "1" ] ; then sleep 5 ; fi ;;
     24) echo "$(tput setaf 1)PF-build.sh stopped due to compiling errors! Try to restore modified files.$(tput sgr0)"; check_script_failed_nr1 ; check_script_failed_nr2 ; cleanup_firmware ; exit 24 ;;
     25) echo "$(tput setaf 1)Failed to execute $(tput sgr0)" ; exit 25 ;;
+    26) echo "$(tput setaf 1)FW_FLAVAVOR or FW_FLAVERSION commented out in 'Firmware/Configuration.h'$(tput sgr0)"; check_script_failed_nr1 ; check_script_failed_nr2 ; cleanup_firmware ; exit 26 ;;
 esac
 }
 #### End: Failures
@@ -543,11 +547,11 @@ fi
 #### Start: Set build environment 
 set_build_env_variables()
 {
-BUILD_ENV="1.0.7"
+BUILD_ENV="1.0.8"
 BOARD="prusa_einsy_rambo"
 BOARD_PACKAGE_NAME="PrusaResearch"
 if [ "$ARDUINO_ENV" == "1.8.19" ]; then
-    BOARD_VERSION="1.0.5-1"
+    BOARD_VERSION="1.0.5-2"
 else
     BOARD_VERSION="1.0.4"
 fi
@@ -926,7 +930,6 @@ fi
 #Check if Build is selected via argument '-b'
 if [ ! -z "$build_flag" ] ; then
     if [[ "$build_flag" == "Auto" && "$git_available" == "1" ]] ; then
-        echo "Build changed to $build_flag"
         BUILD=$(git rev-list --count HEAD)
     elif [[ $build_flag =~ ^[0-9]+$ ]] ; then
         BUILD=$build_flag
@@ -935,27 +938,26 @@ if [ ! -z "$build_flag" ] ; then
         echo "Only $(tput setaf 2)'Auto' (git needed) or numbers $(tput sgr0) are allowed as build '-b' argument!$(tput sgr0)"
         failures 5
     fi
-    echo "New Build number is: $BUILD"
 fi
 
 #Check git branch has changed
 if [ ! -z "git_available" ]; then
-    BRANCH=""
-    CLEAN_PF_FW_BUILD=0
-else
     BRANCH=$(git branch --show-current)
     echo "Current branch is:" $BRANCH
     if [ ! -f "$SCRIPT_PATH/../PF-build.branch" ]; then
-        echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
+        #echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
         echo "created PF-build.branch file"
     else
         PRE_BRANCH=$(cat "$SCRIPT_PATH/../PF-build.branch")
-        echo "Previous branch was:" $PRE_BRANCH
+        #echo "Previous branch was:" $PRE_BRANCH
         if [ ! "$BRANCH" == "$PRE_BRANCH" ] ; then
             CLEAN_PF_FW_BUILD=1
             echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
         fi
     fi
+else
+    BRANCH=""
+    CLEAN_PF_FW_BUILD=0
 fi
 }
 #### End: Getting arguments for command line compiling
@@ -1019,21 +1021,27 @@ prepare_code_for_compiling()
     MOTHERBOARD=$(grep --max-count=1 "\bMOTHERBOARD\b" $SCRIPT_PATH/Firmware/variants/$VARIANT.h | sed -e's/  */ /g' |cut -d ' ' -f3)
     # Check development status
     FW_FLAV=$(grep --max-count=1 "//#define FW_FLAVOR\b" $SCRIPT_PATH/Firmware/Configuration.h|cut -d ' ' -f1)
+    FW_FLAVREV=$(grep --max-count=1 "//#define FW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h|cut -d ' ' -f1)
+    if [[ "$FW_FLAV" == "//#define" && -z $FW_FLAVREV ]];then
+        failures 26
+    fi
     #echo "FLAV:$FW_FLAV"
     if [[ "$FW_FLAV" != "//#define" ]] ; then
         FW_FLAVOR=$(grep --max-count=1 "\bFW_FLAVOR\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
         #echo "FLAVOR:$FW_FLAVOR"
-        FW_FLAVERSION=$(grep --max-count=1 "\bFW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
-        if [[ "$FW_FLAVOR" != "//#define FW_FLAVOR" ]] ; then
+        if [[ "$FW_FLAVREV" != "//#define" ]] ; then
+            FW_FLAVERSION=$(grep --max-count=1 "\bFW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
             FW="$FW-$FW_FLAVOR"
             DEV_CHECK="$FW_FLAVOR"
-            #echo "DEV:$DEV_CHECK"
+            echo "DEV:$DEV_CHECK"
             if [ ! -z "$FW_FLAVERSION" ] ; then
                 FW="$FW$FW_FLAVERSION"
             fi
+        else
+            failures 26
         fi
     fi
-    DEV_CHECK=$(grep --max-count=1 "\bFW_VERSION\b" $SCRIPT_PATH/Firmware/Configuration.h | sed -e's/  */ /g'|cut -d '"' -f2|sed 's/\.//g'|cut -d '-' -f2)
+    #DEV_CHECK=$(grep --max-count=1 "\bFW_VERSION\b" $SCRIPT_PATH/Firmware/Configuration.h | sed -e's/  */ /g'|cut -d '"' -f2|sed 's/\.//g'|cut -d '-' -f2)
     if [ -z "$DEV_STATUS_SELECTED" ] ; then
         if [[ "$DEV_CHECK" == *"RC"* ]] ; then
             DEV_STATUS="RC"
@@ -1157,7 +1165,7 @@ prepare_variant_for_compiling()
     fi
 
     #Prepare Configuration.h to use the correct FW_DEV_VERSION to prevent LCD messages when connecting with OctoPrint
-    sed -i -- "s/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/g" $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- "s/#define FW_DEV_VERSION FW_VERSION_.*/#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/g" $SCRIPT_PATH/Firmware/Configuration.h
 
     # set FW_REPOSITORY
     sed -i -- 's/#define FW_REPOSITORY "Unknown"/#define FW_REPOSITORY "Prusa3d"/g' $SCRIPT_PATH/Firmware/Configuration.h
@@ -1387,7 +1395,9 @@ save_en_firmware()
 cleanup_firmware()
 {
     if [[ -z "$prusa_flag" || "$prusa_flag" == "0" ]]; then
-        rm $SCRIPT_PATH/Firmware/Configuration_prusa.h || failures 13
+        if [ -e "$SCRIPT_PATH/Firmware/Configuration_prusa.h" ]; then
+            rm $SCRIPT_PATH/Firmware/Configuration_prusa.h || failures 13
+        fi
     fi
     # Delete dupblicates
     if find $SCRIPT_PATH/lang/ -name '*RAMBo10a*.txt' -printf 1 -quit | grep -q 1
@@ -1414,8 +1424,8 @@ cleanup_firmware()
     fi
 
     # Restore files to previous state
-    sed -i -- "s/^#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/g" $SCRIPT_PATH/Firmware/Configuration.h
-    sed -i -- 's/^#define FW_REPOSITORY "Prusa3d"/#define FW_REPOSITORY "Unknown"/g' $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- "s/^#define FW_DEV_VERSION FW_VERSION_.*/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/g" $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- 's/^#define FW_REPOSITORY.*/#define FW_REPOSITORY "Unknown"/g' $SCRIPT_PATH/Firmware/Configuration.h
     if [ ! -z "$BUILD_ORG" ] ; then
         sed -i -- "s/^#define FW_COMMIT_NR.*/#define FW_COMMIT_NR $BUILD_ORG/g" $SCRIPT_PATH/Firmware/Configuration.h
     fi
@@ -1439,11 +1449,11 @@ cleanup_firmware()
     fi
 
     # Restore build env files to previous state
-    if [ $BOARD_MEM_MODIFIED == "1" ]; then
+    if [ "$BOARD_MEM_MODIFIED" == "1" ]; then
         sed -i -- "s/^#define FLASHEND .*$/#define FLASHEND        0x3FFFF/g" $BUILD_ENV_PATH/hardware/tools/avr/avr/include/avr/iom2560.h
         echo "$(tput setaf 2)Restored Board Mem$(tput sgr 0)"
     fi
-    if [ $BOARD_FLASH_MODIFIED == "1" ]; then
+    if [ "$BOARD_FLASH_MODIFIED" == "1" ]; then
         sed -i -- "s/^prusa_einsy_rambo.upload.maximum_size.*/prusa_einsy_rambo.upload.maximum_size=253952/g" $BUILD_ENV_PATH/portable/packages/$BOARD_PACKAGE_NAME/hardware/avr/$BOARD_VERSION/boards.txt
         sed -i -- "s/^#define RAMEND.*/#define RAMEND          0x21FF/g" $BUILD_ENV_PATH/hardware/tools/avr/avr/include/avr/iom2560.h
         echo "$(tput setaf 2)Restored Board Flash$(tput sgr 0)"
